Program to Add Two Integers.c
In this program, the user is asked to enter two integers. These two integers are stored in variables number1 and number2 respectively.
Then, these two numbers are added using the + operator, and the result is stored in the sum variable.
Finally, the printf() function is used to display the sum of numbers.