Program Using the long keyword.c
In this program, the sizeof operator is used to find the size of int, long, long long, double and long double variables.
The size of long int and long double variables are larger than int and double variables, respectively.
The sizeof operator returns size_t (unsigned integral type).
The size_t data type is used to represent the size of an object. The format specifier used for size_t is %zu.