Program to Check Even or Odd.c
In the program, the integer entered by the user is stored in the variable num.
Then, whether num is perfectly divisible by 2 or not is checked using the modulus % operator.
If the number is perfectly divisible by 2, test expression number%2 == 0 evaluates to 1 (true). This means the number is even.
However, if the test expression evaluates to 0 (false), the number is odd.